Hypixel-PvP-Tracker-master/pyproject.toml
[tool.poetry]
name = "hypixel-pvp-tracker"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
gevent = "^21.1.2"
hypixel = "^0.7.9"
pymongo = "^3.11.3"
dblpy = "^0.4.0"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
